Some fixes

+++ db/migrations/20170923043553_create_todos.rb
@@ -4,7 +4,7 @@ Hanami::Model.migration do
       primary_key :id
       column :user_id, Integer, null: false
       column :title, String, null: false
-      column :progress, Float, null: false
+      column :progress, Integer, null: false
 
       column :created_at, DateTime, null: false
       column :updated_at, DateTime, null: false
